backend/requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
ffmpeg-python
openai-whisper
opencv-python-headless
ultralytics
Pillow
requests
deep-translator
python-dotenv
